Breaking catalog trait in 2
