+++ iot-edge-spiffe-server/admin-api/src/http/create_get_update_delete_entries.rs
@@ -6,7 +6,7 @@
 use std::borrow::Cow;
 
 use crate::{uri, Api};
-use catalog::Catalog;
+use catalog::Entries;
 use http::{Extensions, StatusCode};
 use http_common::{server, DynRangeBounds};
 use server_admin_api::{
@@ -14,7 +14,10 @@ use server_admin_api::{
     update_registration_entries, ApiVersion,
 };
 
-pub(super) struct Route<C: Catalog + Send + Sync> {
+pub(super) struct Route<C>
+where
+    C: Entries + Send + Sync + 'static,
+{
     page_size: Option<String>,
     page_token: Option<String>,
     api: Api<C>,
@@ -23,7 +26,7 @@ pub(super) struct Route<C: Catalog + Send + Sync> {
 #[async_trait::async_trait]
 impl<C> server::Route for Route<C>
 where
-    C: Catalog + Send + Sync,
+    C: Entries + Send + Sync + 'static,
 {
     type ApiVersion = ApiVersion;
     type Service = super::Service<C>;
