+++ iot-edge-spiffe-server/admin-api/src/http/mod.rs
@@ -1,20 +1,23 @@
 // Copyright (c) Microsoft. All rights reserved.
 
 use crate::Api;
-use catalog::Catalog;
+use catalog::Entries;
 use http_common::make_service;
 use server_admin_api::ApiVersion;
 
 mod create_get_update_delete_entries;
 mod get_select_entries;
 
-pub struct Service<C: Catalog + Send + Sync> {
+pub struct Service<C>
+where
+    C: Entries + Send + Sync + 'static,
+{
     pub(crate) api: Api<C>,
 }
 
 impl<C> Clone for Service<C>
 where
-    C: Catalog + Send + Sync,
+    C: Entries + Send + Sync + 'static,
 {
     fn clone(&self) -> Self {
         Self {
@@ -26,7 +29,7 @@ where
 make_service! {
     service: Service<C>,
     {<C: 'static>}
-    {C:Catalog + Sync + Send}
+    {C: Entries + Send + Sync + 'static}
     api_version: ApiVersion,
     routes: [
         create_get_update_delete_entries::Route<C>,
